+++ src/scheduled-jobs/common/ReconcileEnrollment.js
@@ -1,0 +1,8 @@
+'use strict';
+
+// Export scheduled function
+module.exports = async function() {
+	this.logger.info(`Starting enrollment reconciliation job`, { college: this.name });
+	await this.reconcileEnrollment();
+	this.logger.info(`Completed enrollment reconciliation job`, { college: this.name });
+};